feat: Add dark mode to site and sidebar
This commit adds a new dark mode option to the site and sidebar components. The dark mode is toggled using a button in the top-right corner of the page, which switches the background color of the site and sidebar to a darker color scheme.

By adding a dark mode option, we provide a more customizable user experience and enable users to better tailor the application to their preferences.